Change to admin_s directory
