+++ app/api/api.rb
@@ -22,5 +22,7 @@ class API < Grape::API
   mount V1::Users
   mount V1::Registrations
   mount V1::Confirmations
-  mount V1::Patients
+  namespace :admin do
+    mount V1::Admin::Patients
+  end
 end
